backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
langchain
langchain-ollama
langchain-community
langchain-chroma
langgraph
python-dotenv
fpdf
httpx
pydantic